ai-service/requirements.txt
fastapi
uvicorn
scikit-learn
python-docx
PyPDF2
python-multipart